requirements.txt
seleniumbase
pandas
pyyaml
pytest
